Moved depsgraph_objects() to export/object.py. Created InstanceData
