+++ src/hdusd/engine/engine.py
@@ -20,7 +20,6 @@ import bpy
 from pxr import UsdGeom
 
 from ..utils.stage_cache import CachedStage
-from ..utils import depsgraph_objects
 from ..export import object, world
 
 from ..utils import logging
@@ -50,9 +49,9 @@ class Engine:
         root_prim = stage.GetPseudoRoot()
 
         objects_len = len(depsgraph.objects)
-        for i, obj in enumerate(depsgraph_objects(depsgraph,
-                                                  space_data=space_data,
-                                                  use_scene_lights=use_scene_lights)):
+        for i, obj in enumerate(object.depsgraph_objects(depsgraph,
+                                                         space_data=space_data,
+                                                         use_scene_lights=use_scene_lights)):
             if test_break and test_break():
                 return None
