+++ src/hdusd/utils/__init__.py
@@ -108,35 +108,6 @@ def time_str(val):
     return f"{math.floor(val / 60):02}:{math.floor(val % 60):02}.{math.floor((val % 1) * 100):02}"
 
 
-def depsgraph_objects(depsgraph, *, space_data=None, use_scene_lights=True):
-    SUPPORTED_TYPES = ('MESH', 'LIGHT', 'CURVE', 'FONT', 'SURFACE', 'META', 'CAMERA')
-
-    for obj in depsgraph.objects:
-        if obj.type not in SUPPORTED_TYPES:
-            continue
-
-        if obj.type == 'LIGHT' and not use_scene_lights:
-            continue
-
-        if space_data and not obj.visible_in_viewport_get(space_data):
-            continue
-
-        yield obj
-
-    for instance in depsgraph.object_instances:
-        obj = instance.object
-        if obj.type not in SUPPORTED_TYPES:
-            continue
-
-        if obj.type == 'LIGHT' and not use_scene_lights:
-            continue
-
-        # if space_data and not instance.visible_in_viewport_get(space_data):
-        #     continue
-
-        yield instance
-
-
 def title_str(str):
     s = str.replace('_', ' ')
     return s[:1].upper() + s[1:]
